requirements.txt
django==3.2.8
djangorestframework==3.12.4

#This is for postgres->pip install django psycopg2==2.9.1